+++ docs/julia/heater_3d.jl
@@ -11,7 +11,7 @@
 #   kernelspec:
 #     display_name: base
 #     language: julia
-#     name: julia-1.9
+#     name: julia-1.10
 # ---
 
 # %% [markdown]
@@ -102,7 +102,8 @@ thermal_diffisitivities =
 
 # %% tags=["remove-stderr", "hide-output"]
 boundary_potentials = Dict(["metal3#e1___None" => 0.4, "metal3#e2___None" => 0.0])
-boundary_temperatures = Dict("metal3#e1___None" => 0.4, "metal3#e2___None" => 0.0)
+boundary_temperatures =
+    Dict("metal3#e1___None" => 0.4, "metal3#e2___None" => 0.0, "box___None" => 0.0)
 
 # %% [markdown]
 # Now we're ready to do the simulations! First we simulate the electrical potential,